--- a/package.json
+++ b/package.json
@@ -29,6 +29,8 @@
   "dependencies": {
     "babel-cli": "6.23.0",
     "babel-plugin-transform-object-rest-spread": "6.23.0",
+    "babel-plugin-transform-react-constant-elements": "6.9.1",
+    "babel-plugin-transform-react-inline-elements": "6.8.0",
     "babel-preset-es2015": "6.22.0",
     "babel-preset-react": "6.23.0",
     "chokidar": "1.6.1",