--- a/prod/babel.json
+++ b/prod/babel.json
@@ -1,3 +1,4 @@
 {
-  "presets": [["es2015", {"modules": false}], "react"]
+  "presets": [["es2015", {"modules": false}], "react"],
+  "plugins": ["transform-react-constant-elements", "transform-react-inline-elements"]
 }