Benchmark results:

```diff
-ttfb: 2255 ms
+ttfb: 2014 ms

-total: 2266 ms
+total: 2017ms
````